chore(modal): Move assistive focus into a common hook
