+++ modules/modal/react/lib/ModalContent.tsx
@@ -6,6 +6,7 @@ import Popup, {
   PopupPadding,
   usePopupStack,
   useCloseOnEscape,
+  useAssistiveHideSiblings,
   useFocusTrap,
 } from '@workday/canvas-kit-react-popup';
 
@@ -139,35 +140,6 @@ const useInitialFocus = (
       }
     };
   }, [modalRef, firstFocusRef]);
-};
-
-/**
- * Hides all siblings from assistive technology. Very useful for modal dialogs.
- * This will set `aria-hidden` for siblings of the provided ref and restore
- * the previous `aria-hidden` to each component when the component is unmounted.
- * @param ref React.Ref of the element
- */
-const useAssistiveHideSiblings = <E extends HTMLElement>(ref: React.RefObject<E>) => {
-  React.useEffect(() => {
-    const siblings = [...((ref.current?.parentElement?.children as any) as HTMLElement[])].filter(
-      el => el !== ref.current
-    );
-    const prevAriaHidden = siblings.map(el => el.getAttribute('aria-hidden'));
-    siblings.forEach(el => {
-      el.setAttribute('aria-hidden', 'true');
-    });
-
-    return () => {
-      siblings.forEach((el, index) => {
-        const prev = prevAriaHidden[index];
-        if (prev) {
-          el.setAttribute('aria-hidden', prev);
-        } else {
-          el.removeAttribute('aria-hidden');
-        }
-      });
-    };
-  }, [ref]);
 };
 
 const ModalContent = ({
